pao:hypershift: `ControlPlaneClient`
On hypershift, there is a full separation between the control and data plane.
This means there's a separate client for each.

This commit implements the control plane client.
In order to preserve compatibility between HCP and OCP,
the client encodes/decodes the objects in/from a `ConfigMap` automatically.

This allows the user to perform similar CRUD operations without having the need
to distinguish between HCP and OCP platforms.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>