e2e: pao: refactor mcp.WaitForCondition calls
Refactor mcp.WaitForCondition calls to utilize the newly combined functions (WaitForTuningUpdating, WaitForTuningUpdated) that will support hypershift and OCP.

Note: On OCP, these functions will wait only on the worker-cnf MCP, and not on the worker node as the code did before. Since there was no use of the non-performance worker node in the irqbalance test, I think it is safe to omit that.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>