e2e: added node inspector cleanup on every suite
Currently we are cleaning up the node inspector in the following two ways:
1. Z_deconfig that runs in the end of all suites.
2. We added a trap command in the hack/run-test.sh script to execute the Z_deconfig suite once tests finished or interrupted.

However, when we import some of these suites to an external repo (e.g cnf-features-depoly) none of the cleaning up methods described above are applied.
Therefore we risk that the node inspecor will be left as a side effect after tests complete.
Adding the cleanup after in the end of each suite, will reduce the risk of such a scenario to occur.
If we chose to delete after each suite, we can safely remove the Z_deconfig and the trap command as they are redundant.